APR-020 Envio de e-mail ao mudar status do Pedido

+++ src/main/java/com/atacado/presentes/api/controller/UsuarioController.java
@@ -38,7 +38,7 @@ public class UsuarioController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Usuario> listarPeloId(@PathVariable("id") Integer id) {
+    public ResponseEntity<Usuario> listarPeloId(@PathVariable("id") Long id) {
         Optional<Usuario> usuario = usuarioRepository.findById(id);
 
         if (usuario.isEmpty()) {
@@ -49,7 +49,7 @@ public class UsuarioController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<?> atualizarUsuario(@PathVariable("id") Integer id, @RequestBody Usuario usuario) {
+    public ResponseEntity<?> atualizarUsuario(@PathVariable("id") Long id, @RequestBody Usuario usuario) {
         Optional<Usuario> usuarioExistente = usuarioRepository.findById(id);
 
         if (usuarioExistente.isPresent()) {
@@ -65,7 +65,7 @@ public class UsuarioController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<?> removerPeloId(@PathVariable Integer id) {
+    public ResponseEntity<?> removerPeloId(@PathVariable Long id) {
         Optional<Usuario> usuario = usuarioRepository.findById(id);
 
         if (usuario.isEmpty()) {
